requirements.txt
oauthlib==1.0.3
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
wheel==0.24.0
